Auto-run the installer at the end of the setup process

+++ setup_mac.py
@@ -3,12 +3,6 @@ cd Dropbox/Codes/password_generator
 py setup_mac.py
 '''
 
-# Installing PyInstaller that supports Python 3.6
-'''
-git clone https://github.com/pyinstaller/pyinstaller.git
-cd pyinstaller
-py setup.py install
-'''
 # Dictionary names and keys taken from py2exe settings
 
 dict_console = {
@@ -16,8 +10,9 @@ dict_console = {
     'dest_base': 'Password Generator',
     'icon_resources': [(1, './icons/password_icon.icns')],
     'script': 'password_generator.py',
-    'version': '1.2.6'
+    'version': '1.2.7'
 }
+
 dict_options = {
     'excludes': [
         '_bz2', '_frozen_importlib', '_hashlib', '_lzma', '_ssl',
@@ -39,25 +34,26 @@ def zero_pad(str_ver):
     return str_ver2
 
 
-folder_dist = 'mac'
-
-list_excluded = []
-for lib in dict_options['excludes']:
-    list_excluded.append('--exclude-module')
-    list_excluded.append(lib)
-
-list_pyinstaller = [
-    'pyinstaller', '--onefile',
-    '--distpath', folder_dist,
-    '--windowed',
-    '--icon', dict_console['icon_resources'][0][1],
-    '--name', dict_console['dest_base']
-] + list_excluded + [dict_console['script']]
-
 if __name__ == "__main__":
     import os
     import shutil
     import subprocess
+
+    folder_dist = 'mac'
+
+    list_excluded = []
+    for lib in dict_options['excludes']:
+        list_excluded.append('--exclude-module')
+        list_excluded.append(lib)
+
+    list_pyinstaller = [
+        'python3', '-m',
+        'PyInstaller', '--onefile',
+        '--distpath', folder_dist,
+        '--windowed',
+        '--icon', dict_console['icon_resources'][0][1],
+        '--name', dict_console['dest_base']
+    ] + list_excluded + [dict_console['script']]
 
     if os.path.exists(folder_dist):
         shutil.rmtree(folder_dist)
@@ -66,8 +62,12 @@ if __name__ == "__main__":
     subprocess.run(list_pyinstaller)
 
     print_with_border('Cleaning')
-    shutil.rmtree('__pycache__')
-    shutil.rmtree('build')
-    os.remove(''.join([dict_console['dest_base'], '.spec']))
+    if os.path.exists('__pycache__'):
+        shutil.rmtree('__pycache__')
+    if os.path.exists('build'):
+        shutil.rmtree('build')
+    if os.path.exists(''.join([dict_console['dest_base'], '.spec'])):
+        os.remove(''.join([dict_console['dest_base'], '.spec']))
 
-    print('Executable and installer for v' + dict_console['version'], 'created.')
+    print('Executable for v' + dict_console['version'], 'created.')
+    subprocess.call(['open', 'mac'])
